FirebaseTenantAwareAuth
