+++ FirebaseAdmin/FirebaseAdmin/Auth/FirebaseTenantAwareAuth.cs
@@ -1,0 +1,30 @@
+// Copyright 2018, Google Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+
+namespace FirebaseAdmin.Auth
+{
+    /// <summary>
+    /// The tenant aware <see cref="FirebaseAuth"/> class. You can
+    /// get an instance of this class via <c>FirebaseAuth.DefaultInstance.TenantManager.AuthForTenant</c>.
+    /// </summary>
+    public sealed class FirebaseTenantAwareAuth : BaseAuth
+    {
+        internal FirebaseTenantAwareAuth(FirebaseAuthArgs args)
+            : base(args)
+        {
+        }
+    }
+}
